requirements.txt
mysql-connector-python
nltk
PyPDF2
